v2,0
A different approach has been used for version 2.0. This time instead of another function to select request, changes were made to the selectGrant function itself. The selectGrant function selects the grant for each output by considering not only its own grant pointer but also the input's last accepted output (which is the same as the request pointer). The function has two nested for-loops: the outer loop iterates the grant pointer through all inputs while the inner loop iterates the input's request pointer through all of its VOQs. The function checks if the VOQ pointed to by the combination of these two pointers has packets or not. Based on this, it either grants to the input or continues iterating the pointers. This ensures that all inputs receive only one grant as intended in the original DRRM algorithm.